fix final formatting
