+++ BackendBench/utils.py
@@ -227,3 +227,39 @@ def cleanup_memory_and_gpu():
     gc.collect()
     torch.cuda.synchronize()
     torch.cuda.empty_cache()
+
+
+def clean_op_name_for_directory(op_name: str) -> str:
+    """Convert operator name to valid directory name.
+
+    Examples:
+    - aten::add.Tensor -> add
+    - aten::add.out -> add_out
+    - aten::native_batch_norm -> native_batch_norm
+    - torch.ops.aten.add.default -> add
+    """
+    # Remove aten:: prefix
+    if op_name.startswith("aten::"):
+        op_name = op_name[6:]
+
+    # Remove torch.ops.aten. prefix
+    if op_name.startswith("torch.ops.aten."):
+        op_name = op_name[15:]
+
+    # Handle .default, .Tensor, .out suffixes
+    if "." in op_name:
+        parts = op_name.split(".")
+        base = parts[0]
+        suffix = parts[1] if len(parts) > 1 else ""
+
+        # For common suffixes, we might want to keep them to distinguish overloads
+        if suffix in ["out", "inplace", "scalar"]:
+            op_name = f"{base}_{suffix}"
+        else:
+            # For .default, .Tensor, etc., just use the base name
+            op_name = base
+
+    # Replace any remaining invalid characters
+    op_name = op_name.replace(":", "_").replace("/", "_").replace("\\", "_")
+
+    return op_name
